New version of the box with spherical end
The box now has spherical end rather than flat one. It increases the volume, but most important it increases the strength of the box.
The fixation ring is now flat rather than having a slope.
This part is meant to be printed with the round part up.
It is compatible with the previous lid and gasket. One could use two of this part for a bigger box, with the gasket in between.
Later on i'll release a version of the gasket with a central hole so that it makes a single chamber rather than two.